+++ src/Piece/PieceCreate/PieceCreate.js
@@ -1,0 +1,3 @@
+export const PieceCreate = () => {
+  return <>Create</>;
+};