+++ src/Piece/PieceCreate/PieceCreateContainer.js
@@ -1,0 +1,3 @@
+export const PieceCreateContainer = () => {
+  return <div>PieceCreateContainer</div>;
+};