+++ src/Piece/PieceCreate/index.js
@@ -1,0 +1,1 @@
+export { PieceCreateContainer } from './PieceCreateContainer';