+++ src/Piece/PieceEdit/PieceEdit.js
@@ -1,0 +1,3 @@
+export const PieceEdit = () => {
+  return <>Edit</>;
+};