+++ src/Piece/PieceEdit/index.js
@@ -1,0 +1,1 @@
+export { PieceEditContainer } from './PieceEditContainer';