+++ src/Piece/PieceForm/PieceForm.js
@@ -1,0 +1,3 @@
+export const PieceForm = () => {
+  return <>Hi</>;
+};