+++ src/Piece/index.js
@@ -1,0 +1,3 @@
+export { PieceCreateContainer } from './PieceCreate';
+export { PieceEditContainer } from './PieceEdit';
+export { PieceForm } from './PieceForm';