+++ src/TitleDetails/TitleDetailsActions/TitleDetailsExpectedActions.js
@@ -21,7 +21,7 @@ import {
 export function TitleDetailsExpectedActions({
   applyFilters,
   filters,
-  openAddPieceModal,
+  onPieceCreate,
   openReceiveList,
   hasReceive,
   disabled,
@@ -49,7 +49,7 @@ export function TitleDetailsExpectedActions({
             data-testid="add-piece-button"
             data-test-add-piece-button
             buttonStyle="dropdownItem"
-            onClick={openAddPieceModal}
+            onClick={onPieceCreate}
             disabled={!canAddPiece}
           >
             <Icon size="small" icon="plus-sign">
@@ -98,7 +98,7 @@ TitleDetailsExpectedActions.propTypes = {
   applyFilters: PropTypes.func.isRequired,
   filters: PropTypes.object.isRequired,
   hasReceive: PropTypes.bool.isRequired,
-  openAddPieceModal: PropTypes.func.isRequired,
+  onPieceCreate: PropTypes.func.isRequired,
   openReceiveList: PropTypes.func.isRequired,
   toggleColumn: PropTypes.func.isRequired,
   visibleColumns: PropTypes.arrayOf(PropTypes.string).isRequired,